# docs/openapi-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Community Feed API",
    "description": "Community-governed Bluesky feed generator where subscribers democratically vote on ranking algorithm weights. Built on AT Protocol.\n\n## Authentication\n- **Governance endpoints** require a session cookie or bearer token from `POST /api/governance/auth/login`.\n- **Admin endpoints** additionally require the caller's DID to be in the `BOT_ADMIN_DIDS` allowlist.\n- **Feed endpoints** are public (called by the Bluesky app). Auth is optional for subscriber tracking.\n- **Transparency endpoints** are public and unauthenticated.",
    "version": "1.1.0",
    "contact": {
      "name": "Community Feed",
      "url": "https://github.com/andrewnordstrom-eng/bluesky-community-feed"
    },
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "governance_session",
        "description": "Session cookie from POST /api/governance/auth/login"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from POST /api/governance/auth/login (for CLI/MCP)"
      }
    },
...
  },
  "paths": {
...
    "/xrpc/app.bsky.feed.sendInteractions": {
      "post": {
        "summary": "Send interaction signals",
        "tags": ["Feed"],
        "description": "Receives user interaction signals (requestMore, requestLess, etc.) from Bluesky clients. Requires a valid JWT in the Authorization header. Stored for analytics and scoring feedback.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "interactions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "item": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 512,
                          "pattern": "^at\\:\\/\\/"
                        },
                        "event": {
                          "type": "string",
                          "pattern": "^app\\.bsky\\.feed\\.defs#[a-zA-Z][a-zA-Z0-9]{1,63}$"
                        },
                        "feedContext": {
                          "type": "string",
                          "maxLength": 512
                        }
                      },
                      "required": ["item", "event"],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 100
                  }
                },
                "required": ["interactions"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empty object on success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Empty object on success"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "TooManyRequests"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryAfterSeconds": {
                      "type": "integer",
                      "description": "Seconds to wait before retrying"
                    },
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
    "/api/governance/auth/login": {
      "post": {
        "summary": "Login with Bluesky credentials",
        "tags": ["Auth"],
        "description": "Authenticate using a Bluesky handle and app password. Returns a session cookie (HttpOnly) and bearer token. Use the cookie for browser-based governance actions, or the bearer token for CLI/MCP clients.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "handle": {
                    "type": "string",
                    "minLength": 1
                  },
                  "appPassword": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["handle", "appPassword"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "did": {
                      "type": "string",
                      "description": "Authenticated user DID",
                      "example": "did:plc:abc123"
                    },
                    "handle": {
                      "type": "string",
                      "description": "Bluesky handle",
                      "example": "alice.bsky.social"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Session expiration timestamp"
                    }
                  },
                  "required": ["success", "did", "handle", "expiresAt"]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "TooManyRequests"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryAfterSeconds": {
                      "type": "integer",
                      "description": "Seconds to wait before retrying"
                    },
                    "correlationId": {
                      "type": "string"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/governance/auth/logout": {
      "post": {
        "summary": "Logout and invalidate session",
        "tags": ["Auth"],
        "description": "Invalidates the current session and clears the session cookie.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Session invalidated"
                    }
                  },
                  "required": ["success", "message"]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": ["error", "message"]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://feed.yourdomain.com",
      "description": "Production"
    }
  ],
...
}